// packages/Components/React/src/types.ts
export interface ComponentProps extends RenderableProps<any> {
	className?: string;
	disableStyles?: boolean;
	style?: string | Record<string, string>;
	theme?: Theme;
}
// ...
export interface BaseFacetValue {
	filtered: boolean;
	label: string;
	count?: number;
	url?: any;
}

export interface ValueFacetValue extends BaseFacetValue {
	value: string;
}